Add files via upload
Added the Cutoff function in the signal processing module to calculate the cutoff frequency before applying the high pass filter to the raw data.  The cutoff frequency is selected to keep the background oscillations but remove a maximum of the DC and low frequency components that polute the Hilbert transform results.