Add files via upload
Added the Sensitivity and Sensitivity2D functions to detect the best range of operation of a SET.  These functions calculate the envelope of the derivative of the signal passed as an argument.

The HighPassFilter and HighPassFilter2D are now called PassFilter(2D) and can be used to apply low pass filters on the signals passed to them.  _butter_highpass also becomes _butter_pass because it is involved in the low pass filters.

The Threshold function is now private and is called _threshold.